NEW - bug 280977: Fix coordinate issues with mouse events and tools https://bugs.eclipse.org/bugs/show_bug.cgi?id=280977

+++ org.eclipse.draw3d.graphics3d/src/java/org/eclipse/draw3d/graphics3d/Graphics3DDraw.java
@@ -472,7 +472,7 @@ public interface Graphics3DDraw {
 
 	public abstract void glLineWidth(float width);
 
-	public abstract void glColor3f(int red, int green, int blue);
+	public abstract void glColor3f(float red, float green, float blue);
 
 	public abstract void glFinish();
