+++ org.eclipse.draw3d.lwjgl/src/java/org/eclipse/draw3d/graphics3d/lwjgl/Graphics3DLwjgl.java
@@ -297,10 +297,10 @@ public class Graphics3DLwjgl implements Graphics3D {
 	/**
 	 * {@inheritDoc}
 	 * 
-	 * @see org.eclipse.draw3d.graphics3d.Graphics3DDraw#glColor3f(int, int,
-	 *      int)
+	 * @see org.eclipse.draw3d.graphics3d.Graphics3DDraw#glColor3f(float, float,
+	 *      float)
 	 */
-	public void glColor3f(int red, int green, int blue) {
+	public void glColor3f(float red, float green, float blue) {
 		org.lwjgl.opengl.GL11.glColor3f(red, green, blue);
 	}
